Move content of sugar.
